Edit Reminder

+++ app/src/main/java/com/example/goal_garden_project/models/Goal.kt
@@ -29,8 +29,8 @@ data class Goal(
     var date: Long,      //save as milliseconds and convert it then to date
     var isFulfilled: Boolean = false,
     var reminderOn: Boolean = false,
-    var reminderTime: Int =0,
-    var reminderInterval: Int =0, //in days
+    var reminderTime: Long =0,
+    var reminderInterval: Long =0, //in days
     var isSeeded: Boolean = true
 )
 //later we could add parameter to change color of pot or form or whatsoever..
